[#47331] parsers: Initial reorganization of RGB parsers
